// frontend/src/api/profile/models/ApiUpdateProfile.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ApiUpdateProfile = {
    /**
     * ФИО пользователя
     */
    description?: string;
    /**
     * Номер телефона
     */
    phone?: string;
    /**
     * Email пользователя
     */
    email?: string;
    /**
     * Дата рождения пользователя
     */
    birthdate?: string;
    /**
     * Пол пользователя
     */
    gender?: string;
    /**
     * Ссылка на аватар пользователя
     */
    avatar?: string;
    /**
     * Ссылка на резюме пользователя
     */
    cv?: string;
};